♻️ 응답 DTO count 관련 필드 Integer -> Long으로 타입 변경

+++ BE/src/main/java/team29/hoorry/issuetracker/core/issue/dto/response/IssuesResponse.java
@@ -12,10 +12,10 @@ import team29.hoorry.issuetracker.core.member.dto.MemberResponse;
 @AllArgsConstructor
 public class IssuesResponse {
 
-	private Integer openIssueCount;
-	private Integer closedIssueCount;
-	private Integer labelCount;
-	private Integer milestoneCount;
+	private Long openIssueCount;
+	private Long closedIssueCount;
+	private Long labelCount;
+	private Long milestoneCount;
 
 	private List<MemberResponse> assignees;
 	private List<IssueLabelResponse> labels;
